// Pharmacy Portal/src/app/models/order.model.ts
// order obj
export interface Order{
    id?:string;
    createdAt?: string;
    liveTrack?: string;
    location?: any;
    orderStatus?: string;
    prescriptionType?: string;
    rating?: string;
    referecnceId?: string;
    updatedAt?: string;
    patientId ?: string;
    patientName?: string;
    contactNum?: string;
    docName?: string;
    docId?: string;
    prescription?: string;
    amount?: number;
}